# docs/swagger.yaml
definitions:
...
  clientError.ErrorResponse:
    properties:
      error:
        type: string
      error_code:
        type: string
    type: object
  db.EntryPosition:
    enum:
    - long
    - short
    type: string
    x-enum-varnames:
    - EntryPositionLong
    - EntryPositionShort
  db.PostCryptoStatus:
    enum:
    - pending
    - fail
    - success
    type: string
    x-enum-varnames:
    - PostCryptoStatusPending
    - PostCryptoStatusFail
    - PostCryptoStatusSuccess
  model.ChangePasswordRequest:
    properties:
      new:
        type: string
      old:
        type: string
    required:
    - new
    - old
    type: object
  model.ChangeUsernameRequest:
    properties:
      new:
        type: string
    required:
    - new
    type: object
  model.ChangeUsernameResponse:
    properties:
      token:
        type: string
    type: object
  model.CheckUsernameResponse:
    properties:
      validity:
        type: boolean
    type: object
  model.CreateFeedbackRequest:
    properties:
      message:
        type: string
      type:
        description: optional
        type: string
    required:
    - message
    type: object
  model.CreateMoniestRequest:
    properties:
      bio:
        description: optional
        type: string
      card_id:
        type: string
      description:
        description: optional
        type: string
      fee:
        type: number
      message:
        description: optional
        type: string
    required:
    - card_id
    - fee
    type: object
  model.CreatePostRequest:
    properties:
      currency:
        type: string
      description:
        type: string
      direction:
        type: string
      duration:
        type: string
      stop:
        type: number
      target1:
        type: number
      target2:
        type: number
      target3:
        type: number
    required:
    - currency
    - direction
    - duration
    - stop
    - target1
    - target2
    - target3
    type: object
  model.CreatePostResponse:
    properties:
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      direction:
        $ref: '#/definitions/db.EntryPosition'
      duration:
        type: string
      id:
        type: string
      moniest_id:
        type: string
      score:
        type: number
      start_price:
        type: number
      stop:
        type: number
      target1:
        type: number
      target2:
        type: number
      target3:
        type: number
      updated_at:
        type: string
    type: object
  model.Currency:
    properties:
      currency:
        type: string
      price:
        type: string
    type: object
...
  model.GetContentPostResponse:
    properties:
      created_at:
        type: string
      currency:
        type: string
      direction:
        $ref: '#/definitions/db.EntryPosition'
      duration:
        type: string
      finished:
        type: boolean
      id:
        type: string
      start_price:
        type: number
      status:
        $ref: '#/definitions/db.PostCryptoStatus'
      stop:
        type: number
      target1:
        type: number
      target2:
        type: number
      target3:
        type: number
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  model.LoginRequest:
    properties:
      identifier:
        minLength: 1
        type: string
      password:
        minLength: 6
        type: string
    required:
    - identifier
    - password
    type: object
  model.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.OwnUser'
    type: object
  model.Moniest:
    properties:
      bio:
        type: string
      description:
        type: string
      id:
        type: string
      score:
        type: number
      subscription_info:
        $ref: '#/definitions/model.SubscriptionInfo'
    type: object
  model.OwnUser:
    properties:
      background_photo_link:
        type: string
      background_photo_thumbnail_link:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      location:
        type: string
      moniest:
        $ref: '#/definitions/model.Moniest'
      name:
        type: string
      profile_photo_link:
        type: string
      profile_photo_thumbnail_link:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.RegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 1
        type: string
      password:
        type: string
      surname:
        minLength: 1
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - surname
    - username
    type: object
  model.RegisterResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.OwnUser'
    type: object
  model.SendResetPasswordEmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  model.SendVerificationEmailRequest:
    properties:
      redirect_url:
        type: string
    required:
    - redirect_url
    type: object
  model.SubscriptionInfo:
    properties:
      fee:
        type: number
      message:
        type: string
      updated_at:
        type: string
    type: object
  model.UpdateMoniestProfileRequest:
    properties:
      bio:
        type: string
      description:
        type: string
      subscription_info:
        $ref: '#/definitions/model.SubscriptionInfo'
    type: object
  model.UpdateUserProfileRequest:
    properties:
      background_photo:
        type: string
      location:
        type: string
      name:
        type: string
      profile_photo:
        type: string
      surname:
        type: string
    type: object
  model.User:
    properties:
      background_photo_link:
        type: string
      background_photo_thumbnail_link:
        type: string
      created_at:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      location:
        type: string
      moniest:
        $ref: '#/definitions/model.Moniest'
      name:
        type: string
      profile_photo_link:
        type: string
      profile_photo_thumbnail_link:
        type: string
      surname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  model.VerifyEmailResponse:
    properties:
      redirect_url:
        type: string
    type: object
  model.VerifyPasswordResetRequest:
    properties:
      new:
        type: string
      token:
        type: string
    required:
    - new
    - token
    type: object
  model.VerifyTokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
info:
  contact: {}
paths:
  /account/email/send_verification_email:
    post:
      consumes:
      - application/json
      description: Email verification email sender
      parameters:
      - description: redirect_url is required
        in: body
        name: SendVerificationEmailBody
        required: true
        schema:
          $ref: '#/definitions/model.SendVerificationEmailRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: email already verified
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Send Verification Email
      tags:
      - Account
  /account/email/verify_email:
    post:
      consumes:
      - application/json
      description: Verify email by token
      parameters:
      - description: token is required
        in: body
        name: VerifyEmailBody
        required: true
        schema:
          $ref: '#/definitions/model.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.VerifyEmailResponse'
        "400":
          description: already verified email
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "403":
          description: expired token
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: token not found | user not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body & token
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      summary: Verify Email
      tags:
      - Account
  /account/login:
    post:
      consumes:
      - application/json
      description: Login with [email & password] OR [username & password]
      parameters:
      - description: identifier can be email OR username
        in: body
        name: LoginBody
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "403":
          description: wrong password
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: email OR username not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /account/password:
    put:
      consumes:
      - application/json
      description: Authenticated user password change
      parameters:
      - description: ' '
        in: body
        name: ChangePasswordBody
        required: true
        schema:
          $ref: '#/definitions/model.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: wrong password
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Change Password
      tags:
      - Password
  /account/password/change_password:
    post:
      consumes:
      - application/json
      description: Unauthenticated verify token & change password
      parameters:
      - description: token & new fiels are required
        in: body
        name: VerifyTokenBody
        required: true
        schema:
          $ref: '#/definitions/model.VerifyPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: token is expired
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: reset token not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      summary: Verify Token & Change Password
      tags:
      - Password
  /account/password/send_email:
    post:
      consumes:
      - application/json
      description: Unauthenticated user send reset password email
      parameters:
      - description: ' '
        in: body
        name: SendEmailBody
        required: true
        schema:
          $ref: '#/definitions/model.SendResetPasswordEmailRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      summary: Send Reset Password Email
      tags:
      - Password
  /account/password/verify_token:
    post:
      consumes:
      - application/json
      description: Unauthenticated verify token
      parameters:
      - description: token is required
        in: body
        name: VerifyTokenBody
        required: true
        schema:
          $ref: '#/definitions/model.VerifyTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: token is expired
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: reset token not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      summary: Verify Token
      tags:
      - Password
  /account/register:
    post:
      consumes:
      - application/json
      description: Register as user
      parameters:
      - description: ' '
        in: body
        name: RegisterBody
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RegisterResponse'
        "403":
          description: wrong password
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      summary: Register
      tags:
      - Auth
  /account/username:
    patch:
      consumes:
      - application/json
      description: Change username of the user
      parameters:
      - description: new is required
        in: body
        name: ChangeUsernameBody
        required: true
        schema:
          $ref: '#/definitions/model.ChangeUsernameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ChangeUsernameResponse'
        "403":
          description: username is already taken
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body or username
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Change Username
      tags:
      - Account
  /account/usernames/:username/check:
    get:
      consumes:
      - application/json
      description: Check username is valid of not
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CheckUsernameResponse'
        "406":
          description: invalid username
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      summary: Check Username
      tags:
      - Account
...
  /content/moniests:
    get:
      consumes:
      - application/json
      description: Get All moniests
      parameters:
      - description: 'default: 10 & max: 50'
        in: query
        name: limit
        required: true
        type: integer
      - description: 'default: 0'
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Get Content Moniests
      tags:
      - Content
  /content/posts:
    get:
      consumes:
      - application/json
      description: Get Posts for timeline & explore
      parameters:
      - description: 'default: true'
        in: query
        name: subscribed
        required: true
        type: boolean
      - description: 'default: true'
        in: query
        name: active
        required: true
        type: boolean
      - description: 'default: 10 & max: 50'
        in: query
        name: limit
        required: true
        type: integer
      - description: 'default: 0'
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GetContentPostResponse'
            type: array
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Get Content Posts
      tags:
      - Content
  /crypto/currencies:
    get:
      consumes:
      - application/json
      description: Search crypto currencies by name
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Currency'
            type: array
        "406":
          description: invalid name
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error & crypto api error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Crypto Currency Search
      tags:
      - Crypto
  /feedback:
    post:
      consumes:
      - application/json
      description: '[AUTH OPTIONAL] Create Feedback as Authenticated or Anonymous'
      parameters:
      - description: type is optional
        in: body
        name: CreateFeedbackBody
        required: true
        schema:
          $ref: '#/definitions/model.CreateFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - '""': []
        bearerAuth: []
      summary: Create Feedback
      tags:
      - Feedback
  /moniests:
    post:
      consumes:
      - application/json
      description: Turn into moniest
      parameters:
      - description: ' '
        in: body
        name: CreateMoniestBody
        required: true
        schema:
          $ref: '#/definitions/model.CreateMoniestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OwnUser'
        "400":
          description: user is already moniest
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "403":
          description: 'forbidden operation: email is not verified'
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: not found user
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Be Moniest
      tags:
      - Moniest
  /moniests/:username/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe to Moniest
      parameters:
      - description: moniest username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: already subscribed
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "403":
          description: subscribe own
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: moniest is not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Subscribe to Moniest
      tags:
      - Moniest
  /moniests/:username/unsubscribe:
    post:
      consumes:
      - application/json
      description: Unsubscribe from Moniest
      parameters:
      - description: moniest username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: user not subscribed
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "403":
          description: unsubscribe own
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: moniest is not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Unsubscribe from Moniest
      tags:
      - Moniest
  /moniests/posts:
    post:
      consumes:
      - application/json
      description: Create Crypto Post
      parameters:
      - description: '`description` is optional'
        in: body
        name: CreatePostBody
        required: true
        schema:
          $ref: '#/definitions/model.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreatePostResponse'
        "400":
          description: user is not moniest
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Create Post
      tags:
      - Post
  /moniests/profile:
    patch:
      consumes:
      - application/json
      description: Update Moniest Profile details
      parameters:
      - description: all fields are optional
        in: body
        name: UpdateMoniestBody
        required: true
        schema:
          $ref: '#/definitions/model.UpdateMoniestProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OwnUser'
        "403":
          description: user is not moniest
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "404":
          description: user is not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body | invalid bio | invalid desc | invalid fee | invalid
            message
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Update Moniest Profile
      tags:
      - Moniest
  /users/:username:
    get:
      consumes:
      - application/json
      description: get user info with username
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '''email'' field will be visible if user request for own account'
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: not any user with this username
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid username
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Get User by Username
      tags:
      - User
  /users/:username/subscriptions:
    get:
      consumes:
      - application/json
      description: Get subscribed Users
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: 'default: 10 & max: 50'
        in: query
        name: limit
        required: true
        type: integer
      - description: 'default: 0'
        in: query
        name: offset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Get Subscriptions
      tags:
      - User
  /users/profile:
    patch:
      consumes:
      - application/json
      description: Update user profile [name, surname, location, profile photo, background
        photo]
      parameters:
      - description: all fields are optional
        in: body
        name: UpdateUserBody
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OwnUser'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "406":
          description: invalid body & data
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/clientError.ErrorResponse'
      security:
      - bearerAuth: []
      summary: Update User Profile
      tags:
      - User
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"